fix(ci): python-coverage-shards — split oversized pre-push fix-pass test file under 1500-line maintainability limit
The pre-push validation fix-pass regression tests added in #407/#408
grew tests/unit/runtime/test_pr_monitor_pre_push_validation_fix_pass.py
to 1749 lines, tripping test_first_party_code_files_stay_under_line_limit
(MAX_FIRST_PARTY_FILE_LINES = 1500). Split the suite into a
test_*_parts/ directory (parts 001/002, 1111 and ~660 lines) following
the existing convention; all 31 tests preserved and passing.
